Fixed some missing extended elements and serialize problem for attributes with namespaces
Some elements were missing the parsing of extended foreign elements. This was added for now on AtomPerson and AtomEntry. It is to be checked which other also require the fix.
ALso fixed is the serialization of attributes with namespace. There was a bug that the namespace wasn't known to the content handler. This is fixed only for those elements derived from the atomCommonAttributes.